Admin Option Added successfully
i add admin option but there some of bugs but it will be sorved

+++ routes/web.php
@@ -1,4 +1,5 @@
 <?php
+
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\AdminUserController;
 use App\Http\Controllers\Auth\LoginController;
@@ -12,66 +13,91 @@ use App\Http\Controllers\CategoryController;
 use App\Http\Controllers\ArticleController;
 use App\Http\Controllers\CommentController;
 use App\Http\Controllers\AboutController;
-use App\Http\Controllers\HomeController;
 use App\Http\Controllers\Admin\AdminController as Admin;
 use App\Http\Controllers\WelcomeController;
+use App\Http\Controllers\HomeController;
 
 // Public routes accessible to all users
 
-Route::get('/home', [HomeController::class, 'index'])->name('home');
 Route::get('/', [WelcomeController::class, 'index'])->name('welcome');
+
+// Redirect authenticated users to the welcome page
+Route::middleware(['auth'])->group(function () {
+    Route::get('/dashboard', function () {
+        return redirect()->route('welcome');
+    })->name('dashboard');
+});
+
+
+
+
+
 
 
 Route::middleware(['auth', 'admin'])->group(function () {
+
+    Route::get('/admin/dashboard', [AdminController::class, 'authorize'])->name('admin.dashboard');
+    Route::get('/admin/dashboard', [AdminController::class, 'dashboard'])->name('admin.dashboard');
+    // User routes
     Route::get('/admin/users', [AdminController::class, 'manageUsers'])->name('admin.users.index');
     Route::get('/admin/users/create', [AdminController::class, 'createUser'])->name('admin.users.create');
     Route::post('/admin/users', [AdminController::class, 'storeUser'])->name('admin.users.store');
     Route::get('/admin/users/{id}/edit', [AdminController::class, 'editUser'])->name('admin.users.edit');
     Route::put('/admin/users/{id}', [AdminController::class, 'updateUser'])->name('admin.users.update');
     Route::delete('/admin/users/{id}', [AdminController::class, 'deleteUser'])->name('admin.users.delete');
+    Route::delete('/admin/users/{id}', [AdminController::class, 'destroyUser'])->name('admin.users.destroy');
 
+
+    // Category routes
     Route::get('/admin/categories', [AdminController::class, 'manageCategories'])->name('admin.categories.index');
-    // Add routes for creating, editing, updating, and deleting categories
+    Route::get('admin/categories/create', [AdminController::class, 'createCategory'])->name('admin.categories.create');
+    Route::post('admin/categories', [AdminController::class, 'store'])->name('admin.categories.store');
+    Route::get('admin/categories/{category}/edit', [AdminController::class, 'editCategory'])->name('admin.categories.edit');
+    Route::put('admin/categories/{category}', [AdminController::class, 'updateCategory'])->name('admin.categories.update');
+    Route::get('admin/categories/{category}', [AdminController::class, 'showCategory'])->name('admin.categories.show');
+    Route::delete('admin/categories/{category}', [AdminController::class, 'destroyCategory'])->name('admin.categories.destroy');
 
-    //Route::get('/admin/articles', [AdminController::class, 'manageArticles'])->name('admin.articles.index');
-    Route::get('admin/articles/create', [ArticleController::class, 'create'])->name('articles.create');
-    Route::post('admin/articles', [ArticleController::class, 'store'])->name('articles.store');
-    Route::get('admin/articles/{article}/edit', [ArticleController::class, 'edit'])->name('admin.articles.edit');
-    Route::get('articles/{article}/edit', [ArticleController::class, 'edit'])->name('articles.edit');
 
-    Route::put('admin/articles/{article}', [ArticleController::class, 'update'])->name('articles.update');
-    Route::get('admin/articles/{article}', [ArticleController::class, 'show'])->name('articles.show');
-    Route::delete('admin/articles/{article}', [ArticleController::class, 'destroy'])->name('articles.destroy');
+
+    // Article routes
+    Route::get('/admin/articles/create', [ArticleController::class, 'create'])->name('admin.articles.create');
+
+    Route::get('/admin/articles/create', [AdminController::class, 'createArticles'])->name('admin.articles.create');
+    Route::get('/admin/articles/create', [ArticleController::class, 'create'])->name('admin.articles.create');
+    Route::get('/admin/articles', [AdminController::class, 'manageArticles'])->name('admin.articles.index');
+    Route::get('/admin/articles/create', [ArticleController::class, 'create'])->name('admin.articles.createForm');
+    Route::post('/admin/articles', [ArticleController::class, 'store'])->name('admin.articles.store');
+    Route::get('/admin/articles/{article}/edit', [ArticleController::class, 'edit'])->name('admin.articles.edit');
+    Route::put('/admin/articles/{article}', [ArticleController::class, 'update'])->name('admin.articles.update');
+    Route::get('/admin/articles/{article}', [ArticleController::class, 'show'])->name('admin.articles.show');
+    Route::delete('/admin/articles/{article}', [ArticleController::class, 'destroy'])->name('admin.articles.destroy');
 });
-    // Add routes for creating, editing, updating, and deleting articles
 
-    // Add routes for other admin actions
-// Add routes for other admin actions
+
+
+
+
+
+
+
+//Route::get('/articles/create', [ArticleController::class, 'create'])->name('articles.create');
 Route::get('/articles', [ArticleController::class, 'index'])->name('articles.index');
 Route::get('/articles/search', [ArticleController::class, 'search'])->name('articles.search');
 Route::post('/articles', [ArticleController::class, 'store'])->name('articles.store');
 Route::get('/articles/{article}', [ArticleController::class, 'show'])->name('articles.show');
-Route::get('/articles/create', [ArticleController::class, 'create'])->name('articles.create');
 Route::get('/articles/{article}/edit', [ArticleController::class, 'edit'])->name('articles.editForm');
-
-
 Route::put('/articles/{article}', [ArticleController::class, 'update'])->name('articles.update');
 Route::delete('/articles/{id}', [ArticleController::class, 'destroy'])->name('articles.destroy');
-
-
 // Other routes...
 Route::get('/articles/search', [ArticleController::class, 'search'])->name('articles.search');
 
-
 // Categories
+Route::get('/', [CategoryController::class, 'index'])->name('home');
 Route::resource('categories', CategoryController::class)->except(['show']);
 Route::get('/categories/{category}', [CategoryController::class, 'show'])->name('categories.show');
-Route::get('/categories/{category}', [CategoryController::class, 'show'])->name('categories.show');
-
 
 //home
 Route::get('/home', [HomeController::class, 'index'])->name('home');
-
 
 //comment
 Route::post('/comments', [CommentController::class, 'store'])->name('comments.store');
@@ -81,12 +107,25 @@ Route::post('/comments', [CommentController::class, 'addComment'])->name('commen
 //About
 Route::get('/about', [AboutController::class, 'index'])->name('about');
 
-Route::middleware([
-    'auth:sanctum',
-    config('jetstream.auth_session'),
-    'verified'
-])->group(function () {
-    Route::get('/dashboard', function () {
-        return view('dashboard');
-    })->name('dashboard');
-});
+// ...
+
+
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+
+// Authentication Routes
+Route::get('login', 'App\Http\Controllers\Auth\LoginController@showLoginForm')->name('login');
+Route::post('login', 'App\Http\Controllers\Auth\LoginController@login');
+Route::post('logout', 'App\Http\Controllers\Auth\LoginController@logout')->name('logout');
+
+// Registration Routes
+Route::get('register', 'App\Http\Controllers\Auth\RegisterController@showRegistrationForm')->name('register');
+Route::post('register', 'App\Http\Controllers\Auth\RegisterController@register');
+
+// Password Reset Routes
+Route::get('password/reset', 'App\Http\Controllers\Auth\ForgotPasswordController@showLinkRequestForm')->name('password.request');
+Route::post('password/email', 'App\Http\Controllers\Auth\ForgotPasswordController@sendResetLinkEmail')->name('password.email');
+Route::get('password/reset/{token}', 'App\Http\Controllers\Auth\ResetPasswordController@showResetForm')->name('password.reset');
+Route::post('password/reset', 'App\Http\Controllers\Auth\ResetPasswordController@reset')->name('password.update');
